Move MetaType/MetaMember to separate files in Internal namespace
