+++ OpenRiaServices.DomainServices.Client/Framework/Silverlight/Data/ChangeSetBuilder.cs
@@ -5,6 +5,7 @@ using System.ComponentModel.DataAnnotations;
 using System.Globalization;
 using System.Linq;
 using System.Reflection;
+using OpenRiaServices.DomainServices.Client.Internal;
 
 namespace OpenRiaServices.DomainServices.Client
 {
