+++ OpenRiaServices.DomainServices.Client/Framework/Silverlight/Data/EntitySet.cs
@@ -12,6 +12,7 @@ using System.Reflection;
 
 #if HAS_COLLECTIONVIEW
 using System.Windows.Data;
+using OpenRiaServices.DomainServices.Client.Internal;
 #endif
 
 namespace OpenRiaServices.DomainServices.Client
