+++ OpenRiaServices.DomainServices.Client/Framework/Silverlight/OpenRiaServices.DomainServices.Client.csproj
@@ -152,6 +152,7 @@
     <Compile Include="Data\Entity.INotifyDataErrorInfo.cs" />
     <Compile Include="Data\EntityQueryable.cs" />
     <Compile Include="Data\IDomainClientFactory.cs" />
+    <Compile Include="Data\Internal\MetaMember.cs" />
     <Compile Include="Data\MergeAttribute.cs" />
     <Compile Include="Data\LoadResult.cs" />
     <Compile Include="Data\ObjectStateUtility.cs" />
@@ -162,7 +163,7 @@
     <Compile Include="Data\ValidationResultCollection.cs" />
     <Compile Include="Data\EntityVisitor.cs" />
     <Compile Include="Data\InvokeArgs.cs" />
-    <Compile Include="Data\MetaType.cs" />
+    <Compile Include="Data\Internal\MetaType.cs" />
     <Compile Include="Data\RoundtripOriginalAttribute.cs" />
     <Compile Include="ExceptionHandlingUtility.cs" />
     <Compile Include="Data\BinaryTypeUtility.cs" />
